feat: unify admin features under /dashboard
